+++ vosk/CHANGELOG.md
@@ -1,5 +1,6 @@
 # (UNRELEASED) 0.1.1
 * Documentation fixes
+* Loosen generic bounds for Recognizer::new_with_grammar
 
 # 0.1.0
 * First release